backends/native: Change uint32_t modifiers to uint64_t
They were erroneously cast to uint32_t. In particular this prevented the
invalid modifier from working as it is equal to a 64-bit -1.

Part-of: <https://gitlab.gnome.org/GNOME/mutter/-/merge_requests/2322>

+++ src/backends/native/meta-drm-buffer-gbm.c
@@ -124,7 +124,7 @@ meta_drm_buffer_gbm_get_offset (MetaDrmBuffer *buffer,
   return gbm_bo_get_offset (buffer_gbm->bo, plane);
 }
 
-static uint32_t
+static uint64_t
 meta_drm_buffer_gbm_get_modifier (MetaDrmBuffer *buffer)
 {
   MetaDrmBufferGbm *buffer_gbm = META_DRM_BUFFER_GBM (buffer);
