+++ src/backends/native/meta-drm-buffer.h
@@ -63,7 +63,7 @@ uint32_t meta_drm_buffer_get_format (MetaDrmBuffer *buffer);
 int meta_drm_buffer_get_offset (MetaDrmBuffer *buffer,
                                 int            plane);
 
-uint32_t meta_drm_buffer_get_modifier (MetaDrmBuffer *buffer);
+uint64_t meta_drm_buffer_get_modifier (MetaDrmBuffer *buffer);
 
 gboolean meta_drm_buffer_supports_fill_timings (MetaDrmBuffer *buffer);
